requirements.txt
anyio==3.5.0
asgiref==3.5.0
click==8.0.4
fastapi==0.75.0
h11==0.13.0
idna==3.3
numpy==1.22.4
pydantic==1.9.0
PyYAML==6.0
sniffio==1.2.0
starlette==0.17.1
titansemail @ git+https://github.com/lakes-legendaries/titans-email.git@279978579191f7c04e1b0ef40cafb10d938f9fef
typing_extensions==4.1.1
uvicorn==0.17.6
